refactor: move pages into components and style auth forms
